Endpoints

+++ src/controllers/MeetingController.js
@@ -1,9 +1,64 @@
-const {User, Meeting, MeetingMember, CandidateVote, Vote} = require('../models')
+const {QueryTypes} = require("sequelize");
+const {User, Meeting, MeetingMember, CandidateVote, Vote, CandidateMeeting} = require('../models')
+const db = require('../models')
+
+async function getUsers(meetingID) {
+    const userID = await MeetingMember.findAll({
+        attributes: ['userID', 'role'],
+        where: {
+            meetingID: meetingID
+        }
+    });
+    const users = [];
+    for (const i in userID) {
+        await users.push(await User.findOne({
+            attributes: ['name', 'id'],
+            where: {
+                id: userID[i].dataValues.userID
+            }
+        }));
+    }
+    return users;
+}
 
 module.exports = {
+    async getMeeting(req, res) {
+        try {
+            const users = await getUsers(req.query.meetingID);
+            const candidateMeeting = await CandidateMeeting.findAll({
+                attributes: ['candidateID', 'start', 'end', 'length'],
+                where: {
+                    meetingID: req.query.meetingID
+                }
+            })
+            let votes = [];
+            let voteInformation = [];
+            for (const i in candidateMeeting) {
+                votes.push(await db.sequelize.query(`SELECT count(userID) as count, candidateID
+                                                     FROM Votes
+                                                     WHERE candidateID = ${candidateMeeting[i].dataValues.candidateID}`, {type: QueryTypes.SELECT}));
+                voteInformation.push(await Vote.findAll({
+                    attributes: ['candidateID', 'userID'],
+                    where: {
+                        candidateID: candidateMeeting[i].dataValues.candidateID
+                    }
+                }));
+            }
+
+            res.send({
+                users,
+                candidateMeeting,
+                votes,
+                voteInformation
+            })
+        } catch (error) {
+            res.status(500).send({
+                error: 'Something went wrong with getting your meeting, please try again later'
+            })
+        }
+    },
     async createGuestMeeting(req, res) {
         try {
-            console.log(req.body);
             const user = await User.create({
                 name: req.body.name
             });
@@ -21,32 +76,41 @@ module.exports = {
             });
             res.send();
         } catch (error) {
-            console.log(error);
             res.status(500).send({
                 error: 'Something went wrong with creating a meeting, please try again later'
             })
         }
     },
+    async getUsers(req, res) {
+        try {
+            const user = getUsers(req.query.meetingID);
+            res.send({
+                user
+            });
+        } catch (error) {
+            res.status(500).send({
+                error: 'Something went wrong with getting users in this meeting, please try again later'
+            })
+        }
+    },
     async addGuestUser(req, res) {
-      try {
-          const user = await User.create({
-             name: req.body.name
-          });
-          await MeetingMember.create({
-             userID: user.id,
-             meetingID: req.body.meetingID
-          });
-          res.status(200).send();
-      } catch (error) {
-          console.log()
-          res.status(500).send({
-              error: 'Something went wrong with adding a user to the meeting, please try again later'
-          })
-      }
+        try {
+            const user = await User.create({
+                name: req.body.name
+            });
+            await MeetingMember.create({
+                userID: user.id,
+                meetingID: req.body.meetingID
+            });
+            res.status(200).send();
+        } catch (error) {
+            res.status(500).send({
+                error: 'Something went wrong with adding a user to the meeting, please try again later'
+            })
+        }
     },
     async editMeetingType(req, res) {
         try {
-            console.log(req.body.pollType)
             await Meeting.update({
                 pollType: req.body.pollType,
             }, {
@@ -54,9 +118,8 @@ module.exports = {
                     meetingID: req.body.meetingID
                 }
             });
-            res.send();
+            res.status(200).send();
         } catch (error) {
-            console.log()
             res.status(500).send({
                 error: 'Something went wrong with editing the meeting type, please try again later'
             })
@@ -72,9 +135,8 @@ module.exports = {
                     meetingID: req.body.meetingID
                 }
             });
-            res.send();
+            res.status(200).send();
         } catch (error) {
-            console.log()
             res.status(500).send({
                 error: 'Something went wrong with editing meeting details, please try again later'
             });
@@ -82,22 +144,22 @@ module.exports = {
     },
     async deleteMeeting(req, res) {
         try {
-
+            console.log(req.body, req);
             const users = await MeetingMember.findAll({
                 attributes: ['userID'],
                 where: {
-                    meetingID: req.body.meetingID
+                    meetingID: req.query.meetingID
                 }
             })
             const candidateID = CandidateVote.findAll({
                 where: {
-                    meetingID: req.body.meetingID
+                    meetingID: req.query.meetingID
                 }
             })
 
             await MeetingMember.destroy({
                 where: {
-                    meetingID: req.body.meetingID
+                    meetingID: req.query.meetingID
                 }
             });
 
@@ -110,21 +172,20 @@ module.exports = {
                         }
                     })
                 }
-                // todo check if user is a guest user, if it is then delete
-                await User.destroy({
-                    where: {
-                        id: users[i].id
-                    }
-                });
+                if (users[i].guest === true)
+                    await User.destroy({
+                        where: {
+                            id: users[i].id
+                        }
+                    });
             }
             await Meeting.destroy({
                 where: {
-                    meetingID: req.body.meetingID
+                    meetingID: req.query.meetingID
                 }
             });
-            res.send("success");
+            res.status.send();
         } catch (error) {
-            console.log(error);
             res.status(500).send({
                 error: 'Something went wrong with deleting the meeting, please try again later'
             });
